Refactor and add home icon
